+++ site sae 301/header.php
@@ -96,18 +96,20 @@
             <span></span>
         </div>
         <nav class="nav-links desktop">
-            <a href="index.php">Accueil</a>
+            <a href="acceuil.php">Accueil</a>
             <a href="adoption.php">Adoption</a>
             <a href="apropos.php">À propos</a>
         </nav>
         <nav class="nav-links mobile">
-            <a href="index.php">Accueil</a>
+            <a href="acceuil.php">Accueil</a>
             <a href="adoption.php">Adoption</a>
             <a href="apropos.php">À propos</a>
         </nav>
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     </header>
 
     <script>
+        src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"
         // JavaScript for responsive hamburger menu
         const hamburger = document.querySelector('.hamburger');
         const mobileNav = document.querySelector('.nav-links.mobile');
